split some code out from `FoundationDowntimeSectionV2`
