+++ src/parser/classic/CombatLogParser.ts
@@ -55,6 +55,7 @@ import BloodFury from 'parser/shared/modules/racials/orc/BloodFury';
 import GiftOfTheNaaru from 'parser/shared/modules/racials/draenei/GiftOfTheNaaru';
 import Stoneform from 'parser/shared/modules/racials/dwarf/Stoneform';
 import { MeleeUptimeAnalyzer } from 'interface/guide/foundation/analyzers/MeleeUptimeAnalyzer';
+import DowntimeDebuffAnalyzer from 'interface/guide/foundation/analyzers/DowntimeDebuffAnalyzer';
 
 class CombatLogParser extends BaseCombatLogParser {
   static defaultModules: DependenciesDefinition = {
@@ -90,6 +91,7 @@ class CombatLogParser extends BaseCombatLogParser {
     statTracker: StatTracker,
     alwaysBeCasting: AlwaysBeCasting,
     meleeUptime: MeleeUptimeAnalyzer,
+    downtimeDebuffs: DowntimeDebuffAnalyzer,
     filteredActiveTime: FilteredActiveTime,
     abilities: Abilities,
     buffs: Auras,
