+++ src/parser/core/CombatLogParser.tsx
@@ -117,6 +117,7 @@ import SkardynsGrace from 'parser/retail/modules/items/thewarwithin/trinkets/Ska
 import MereldarsToll from 'parser/retail/modules/items/thewarwithin/trinkets/MereldarsToll';
 import CirralConcoctory from 'parser/retail/modules/items/thewarwithin/trinkets/CirralConcotory';
 import { MeleeUptimeAnalyzer } from 'interface/guide/foundation/analyzers/MeleeUptimeAnalyzer';
+import DowntimeDebuffAnalyzer from 'interface/guide/foundation/analyzers/DowntimeDebuffAnalyzer';
 // This prints to console anything that the DI has to do
 const debugDependencyInjection = false;
 const MAX_DI_ITERATIONS = 100;
@@ -180,6 +181,7 @@ class CombatLogParser {
     statTracker: StatTracker,
     alwaysBeCasting: AlwaysBeCasting,
     meleeUptime: MeleeUptimeAnalyzer,
+    downtimeDebuffs: DowntimeDebuffAnalyzer,
     filteredActiveTime: FilteredActiveTime,
     abilities: Abilities,
     buffs: Auras,
